lot of stuf

+++ src/pre_processing/functions/embeddings.py
@@ -7,7 +7,7 @@ embedder = SentenceTransformer("all-MiniLM-L6-v2")
 def compute_embedding(doc):
     # tag = embedding
     if "text" not in doc :
-        return np.array([0]*384)
+        return np.array([0]*384).astype(np.float32)
     article = doc["text"]
     article_embedding = embedder.encode(article, convert_to_tensor=False)
     return article_embedding
@@ -18,7 +18,7 @@ def cos_score(embedding):
         "database OR data reusability OR data reuse OR data sharing OR data access OR data privacy OR data protection OR GDPR"
     ]
     query_embedding = embedder.encode(query, convert_to_tensor=False)
-    cos_scores = util.cos_sim(query_embedding, embedding)[0]
+    cos_scores = util.cos_sim(query_embedding.astype(np.float32), embedding)[0]
     return float(cos_scores.abs())
